requirements.txt
numpy==1.19.5
PyQt5==5.15.7
pyqtgraph==0.12.4
pyserial==3.5
